+++ src/test/java/was/utils/FileIoUtilsTest.java
@@ -1,9 +1,8 @@
-package utils;
+package was.utils;
 
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import was.utils.FileIoUtils;
 
 public class FileIoUtilsTest {
     private static final Logger log = LoggerFactory.getLogger(FileIoUtilsTest.class);
