+++ src/test/java/was/utils/HandlebarsTest.java
@@ -1,4 +1,4 @@
-package utils;
+package was.utils;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Template;
