+++ src/test/java/was/utils/IOUtilsTest.java
@@ -1,9 +1,8 @@
-package utils;
+package was.utils;
 
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import was.utils.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
